packages/mcp_servers/chase_travel/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chase-travel-mcp"
version = "0.1.0"
description = "Chase Travel MCP Server for flight search functionality"
authors = [
    { name = "Chase Travel Team" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "httpx>=0.26.0",
    "mcp>=0.1.0"
]
requires-python = ">=3.9"

[project.scripts]
chase-travel-mcp = "chase_travel.server:main"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto" 